yosys: Work on pre-synthesis processing
* synth: fix faulty sv2v sub_per_hash

For some reason the sub_per_hash from sv2v is faulty.
A svase plan to simplify assignments is in the work.
In the meanwhile this patch fixes the problem.

Both functions in sub_per_hash produce a output that
only depends on the Key parameters, so they are not needed.

* yosys: remove old patch interfering with hyperbus

* yosys: re-add proper hyperbus synthesis patch